requirements.txt
asgiref==3.8.1
Django==4.2
django-phonenumber-field==8.0.0
djangorestframework==3.14.0
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
phonenumberslite==8.13.50
psycopg==3.2.3
psycopg2-binary==2.9.7
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
gunicorn==20.1.0
